+++ cmd/api/handler/namespace.go
@@ -422,8 +422,8 @@ func (handler *NamespaceHandler) BlobMetadata(c echo.Context) error {
 type getBlobLogsForNamespace struct {
 	Id         string `param:"id"         validate:"required,hexadecimal,len=56"`
 	Version    byte   `param:"version"`
-	Limit      uint64 `query:"limit"      validate:"omitempty,min=1,max=100"`
-	Offset     uint64 `query:"offset"     validate:"omitempty,min=0"`
+	Limit      int    `query:"limit"      validate:"omitempty,min=1,max=100"`
+	Offset     int    `query:"offset"     validate:"omitempty,min=0"`
 	Sort       string `query:"sort"       validate:"omitempty,oneof=asc desc"`
 	SortBy     string `query:"sort_by"    validate:"omitempty,oneof=time size"`
 	Commitment string `query:"commitment" validate:"omitempty,base64url"`
@@ -495,8 +495,8 @@ func (handler *NamespaceHandler) GetBlobLogs(c echo.Context) error {
 	}
 
 	fltrs := storage.BlobLogFilters{
-		Limit:      int(req.Limit),
-		Offset:     int(req.Offset),
+		Limit:      req.Limit,
+		Offset:     req.Offset,
 		Sort:       pgSort(req.Sort),
 		SortBy:     req.SortBy,
 		Commitment: cm,
